Platform: change kernel configs for 3.11, 3.12, 3.13 and 3.14
By accident these kernel configurations have the "symbols for
debugging/ksymoops" option enabled. Since this is a useful option for
development is isn't for the regular use of the kernel. Due to the fact this
option enlarges the kernel image it does not fit into the NAND kernel
partition anymore.

This change disables all symbols for debugging/ksymoops to make these kernels
fit again into the 2 MiB NAND kernel partition

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>